chore: clean up code

+++ src/controllers/task-controller.ts
@@ -22,20 +22,6 @@ const getTasks = async (req: Request, res: Response) => {
   }
 };
 
-// const getTasksByUserId = async (req: Request, res: Response) => {
-//   try {
-//     const { userId } = req.params;
-//     const result = await TaskService.findTasksByUserId(parseInt(userId));
-//     res.status(200).json(result);
-//     return;
-//   } catch (error) {
-//     console.error("Error fetching tasks from database: ", error);
-//     res
-//       .status(500)
-//       .json({ message: "Error fetching tasks from database: " + error });
-//     return;
-//   }
-// };
 
 const addTask = async (req: Request, res: Response) => {
   try {
@@ -59,13 +45,15 @@ const removeTask = async (req: Request, res: Response) => {
     const task = await TaskService.findTaskById(id);
    
     //@ts-ignore
-    if (req.user.id !== task.userId && req.user.role === "basic") {
+    if (req.user.id !== task.userId && req.user.role === "basic") { // authenticate user and only allow him to delete his own content
       res
         .status(400)
         .json({ message: "You are not authorized to delete this task." });
       return;
     }
-    const result = await TaskService.deleteTask(id);
+    if(req.user.role ==='admin'){ // explicitly allow only admins to delete other users' content
+      const result = await TaskService.deleteTask(id);
+    }
     res
       .status(200)
       .json({ message: `Successfully deleted task with id ${id}` });
